Integracion continua reparada (Closes #83) + Timer en Long

+++ app/src/main/java/com/gracehopper/laserchessapp/gameLogic/move/Move.kt
@@ -12,5 +12,5 @@ data class Move(
     val to: String?,            // Casilla destino (en caso de traslación)
     val destroyed: String?,     // Casilla en la que una pieza ha sido destruida
     val laserPath: String?,     // Ruta del laser
-    val timer: Double?          // Tiempo restante o actualización de timer
+    val timer: Long?          // Tiempo restante o actualización de timer
 )
