+++ app/src/main/java/com/gracehopper/laserchessapp/gameLogic/move/MoveParser.kt
@@ -40,7 +40,7 @@ object MoveParser {
         val timer = timerPart
             .substringAfter("{")
             .substringBefore("}")
-            .toDouble()
+            .toLong()
 
         /**
          * Extraer captura (x...)
